style : Agent view files updated
