+++ app/views/agent/agentHeader.view.php
@@ -52,10 +52,10 @@
                     <li><a href="<?= ROOT ?>/dashboard/preInspection"><button class="btn"><i class="fa-solid fa-file-contract"></i>Pre Inspections</button></a></li>
                     <li><a href="<?= ROOT ?>/dashboard/tasks"><button class="btn"><i class="fa-solid fa-list-check"></i>Tasks</button></a></li>
                     <li><a href="<?= ROOT ?>/dashboard/requestedTasks"><button class="btn"><i class="fa-solid fa-inbox"></i>Requested Tasks</button></a></li>
-                    <li><a href="<?= ROOT ?>/dashboard/repairings"><button class="btn"><i class="fa-solid fa-screwdriver-wrench"></i>Services</button></a></li>
+                    <li><a href="<?= ROOT ?>/dashboard/services"><button class="btn"><i class="fa-solid fa-screwdriver-wrench"></i>Services</button></a></li>
                     <li><a href="<?= ROOT ?>/dashboard/bookings"><button class="btn"><i class="fa-solid fa-house"></i>Bookings</button></a></li>
                     <li><a href="<?= ROOT ?>/dashboard/inventory"><button class="btn"><i class="fa-solid fa-warehouse"></i>Inventory</button></a></li>
-                    <li><a href="<?= ROOT ?>/dashboard/services"><button class="btn" style="padding-right:0px;"><i class="fa-solid fa-users-gear"></i>Manage Providers</button></a></li>
+                    <li><a href="<?= ROOT ?>/dashboard/manageProviders"><button class="btn" style="padding-right:0px;"><i class="fa-solid fa-users-gear"></i>Manage Providers</button></a></li>
                     <li><a href="<?= ROOT ?>/dashboard/profile" data-section="profile"><button class="btn"><i class="fa-solid fa-user"></i>Profile</button></a></li>
                 </ul>
                 <form method="post" id="logout">
