--- app/views/components/repaircards/electricalRepairing.php
@@ -1,14 +1,0 @@
-<a href='<?= ROOT ?>/dashboard/tasks/ongoingtask/electricalrepairing'>
-    <div class="repair-card1">
-        <div class="repair-image">
-            <img src="<?= ROOT ?>/assets/images/repairimages/electricrepairing.png" alt="services">
-        </div>
-        <div class="property-details">
-            <div class="profile-details-items">
-                <div>
-                    <span class="no-underline"><h3>Electrical Repairing</h3></span>
-                </div>
-            </div>
-        </div>
-    </div>
-</a>
