+++ public/assets/css/userview.css
@@ -9,11 +9,9 @@ body {
 .user_view-container {
     display: flex;
     flex-direction: column;
-    height: 100vh;
     background-color: var(--background-color);
     font-family: 'Poppins', Arial, sans-serif;
     position: relative;
-    overflow: hidden;
 }
 
 .header-line {
